georange v1

+++ db/migrate/20130719141358_add_coordinates_and_address_to_activity.rb
@@ -1,7 +1,7 @@
 class AddCoordinatesAndAddressToActivity < ActiveRecord::Migration
   def change
-    add_column :activities, :latitude, :float
-    add_column :activities, :longitude, :float
+    add_column :activities, :latitude, :double
+    add_column :activities, :longitude, :double
     add_column :activities, :street, :string
     add_column :activities, :city, :string
     add_column :activities, :zip_code, :string
